Refactor: 기존 상수들 위치 변경
