+++ src/utils/constants/city.ts
@@ -1,7 +1,3 @@
-export const INIT_LOCATION = { latitude: 37.5665, longitude: 126.978 };
-export const CENTER_LOCATION = { latitude: 36.2, longitude: 127.7 };
-export const FINE_DUST = '미세먼지';
-export const ULTRA_FINE_DUST = '초미세먼지';
 export const CITY_GROUP = [
   { cityName: '서울', cityNumber: 0 },
   { cityName: '부산', cityNumber: 1 },
