+++ src/utils/constants/dust.ts
@@ -1,0 +1,2 @@
+export const FINE_DUST = '미세먼지';
+export const ULTRA_FINE_DUST = '초미세먼지';
