+++ src/utils/constants/location.ts
@@ -1,0 +1,2 @@
+export const INIT_LOCATION = { latitude: 37.5665, longitude: 126.978 };
+export const CENTER_LOCATION = { latitude: 36.2, longitude: 127.7 };
